reorganise and document
