+++ actix-tls/CHANGES.md
@@ -6,6 +6,16 @@
 * Remove redundant `connect::Connection::from_parts` method. [#422]
 * Rename TLS acceptor service future types and hide from docs. [#422]
 * Implement `Error` for `ConnectError`. [#422]
+* Implement `Error` for `TlsError` where both types also implement `Error`. [#422]
+* Rename `accept::native_tls::{NativeTlsAcceptorService => AcceptorService}`. [#422]
+* Make `ConnectAddrsIter` private. [#422]
+* Rename method `connect::Connection::{host => hostname}`. [#422]
+* Rename struct `connect::{Connect => ConnectionInfo}`. [#422]
+* Rename struct `connect::{ConnectServiceFactory => Connector}`. [#422]
+* Rename struct `connect::{ConnectService => ConnectorService}`. [#422]
+* Remove `connect::{new_connector, new_connector_factory, default_connector, default_connector_factory}` methods. [#422]
+* Convert `connect::ResolverService` from enum to struct. [#422]
+* Remove `connect::native_tls::Connector::service` method. [#422]
 
 [#422]: https://github.com/actix/actix-net/pull/422
 
@@ -48,8 +58,7 @@
 * Remove `connect::ssl::openssl::OpensslConnectService`. [#297]
 * Add `connect::ssl::native_tls` module for native tls support. [#295]
 * Rename `accept::{nativetls => native_tls}`. [#295]
-* Remove `connect::TcpConnectService` type. Service caller expecting a `TcpStream` should use
-  `connect::ConnectService` instead and call `Connection<T, TcpStream>::into_parts`. [#299]
+* Remove `connect::TcpConnectService` type. Service caller expecting a `TcpStream` should use `connect::ConnectService` instead and call `Connection<T, TcpStream>::into_parts`. [#299]
 
 [#295]: https://github.com/actix/actix-net/pull/295
 [#296]: https://github.com/actix/actix-net/pull/296
