+++ actix-tls/Cargo.toml
@@ -13,7 +13,8 @@ license = "MIT OR Apache-2.0"
 edition = "2018"
 
 [package.metadata.docs.rs]
-features = ["openssl", "rustls", "native-tls", "accept", "connect", "uri"]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
 
 [lib]
 name = "actix_tls"
@@ -48,10 +49,12 @@ actix-utils = "3.0.0"
 
 derive_more = "0.99.5"
 futures-core = { version = "0.3.7", default-features = false, features = ["alloc"] }
-http = { version = "0.2.3", optional = true }
 log = "0.4"
 pin-project-lite = "0.2.7"
 tokio-util = { version = "0.6.3", default-features = false }
+
+# uri
+http = { version = "0.2.3", optional = true }
 
 # openssl
 tls-openssl = { package = "openssl", version = "0.10.9", optional = true }
