--- actix-tls/src/connect/tls/mod.rs
@@ -1,10 +1,0 @@
-//! TLS Services
-
-#[cfg(feature = "openssl")]
-pub mod openssl;
-
-#[cfg(feature = "rustls")]
-pub mod rustls;
-
-#[cfg(feature = "native-tls")]
-pub mod native_tls;
